requirements.txt
asgiref==3.4.1
Django==3.2.5
django-crispy-forms==1.12.0
django-phonenumber-field==5.2.0
gunicorn==20.1.0
phonenumberslite==8.12.28
Pillow==8.3.1
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.3.0
